[enhancement/5] Rename refactor → index
